expanded histogram line parsing; added functionality for parsing PDF output
According to the Python documentation, os.popen [3] and os.popen3 [4]
are now deprecated. According to the documentation [5], one should use
subprocess.Popen instead. This change was done. As a result, it was also
possible to check the exit code of the identify utility. For example,
if identify is run on a file type it doesn't know how to handle, it will
exit with the following error:

    identify.im6: no decode delegate for this image format

Now, in such cases, an exception is thrown from this Python module if
the exit code is not zero (which by convention means the command ran
succesfully).

The histogram line has a varying format. It varies on the image and
the kind of channels and colors present in it, it also varies on the
ImageMagick version used.
We have seen differences so far between these two versions
- ImageMagick 6.7.7-10 2014-03-06 Q16
- ImageMagick 6.9.2-4 Q16 x86_64 2015-11-27

The regex for parsing the histogram lines has been expanded.
In the case of PDF files, identify produces output for each page of the document.
Because of that, changes were made in toXML() because we now have an
<Image> tag for each page.
This means that now, even for images (png, jpg, dcm), the general
structure of the tree will look like this:

<Images>
    <Image>..</Image>
    <Image>..</Image>
    ..
</Images>

since we want to cover both multi-page and single-page output.

The toIRODS method is also affected by this change.
The output in iRODS format was intended to be used by the msiString2KeyValPair
function [1] [2], and according to the documentation it's a %-separated
key=value pair string. However, because PDFs can have multiple pages, there
will be multiple <Image> nodes, and this will create collisions in the
iRODS key-value output. Currently, the toIRODS method only returns the
metadata for the first page of a PDF.

In addition to this, some basic unit tests were written to check the
properties of the tree.

[1] http://wiki.irods.org/doxygen/key_val_pair_m_s_8c_ac7af139c64d3eb8820b58fae5b48620e.html
[2] https://github.com/irods/irods/blob/a1db0f5defa9a34f72be3be1e4f8ae24965f9187/iRODS/server/re/src/keyValPairMS.cpp#L165
[3] https://docs.python.org/2/library/os.html#os.popen
[4] https://docs.python.org/2/library/os.html#os.popen3
[5] https://docs.python.org/2/library/subprocess.html#replacing-os-popen-os-popen2-os-popen3